test: port final-commit drift journey to shared daemon client
Fix the integration compile regression after combining #978 and #979. Reuse portable private fixture creation and retain the new identity-drift assertion on every platform.
